Setup fetchslice, store, and organise directory structure
