+++ src/components/Query.js
@@ -1,22 +1,13 @@
 import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import {
-  decrement,
-  increment,
-  incrementByAmount,
-  incrementAsync,
-  selectCount,
-} from './counterSlice';
-import styles from './Counter.module.css';
+// import { useSelector, useDispatch } from 'react-redux';
+import styles from '../styles/Query.module.css';
 
-export function Counter() {
-  const count = useSelector(selectCount);
-  const dispatch = useDispatch();
-  const [incrementAmount, setIncrementAmount] = useState('2');
+const Query = () => {
+
 
   return (
     <div>
-      <div className={styles.row}>
+      {/* <div className={styles.row}>
         <button
           className={styles.button}
           aria-label="Increment value"
@@ -54,7 +45,9 @@ export function Counter() {
         >
           Add Async
         </button>
-      </div>
+      </div> */}
     </div>
   );
 }
+
+export default Query;
