+++ src/lib/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit';
+import fetchDataReducer from './slices/fetchDataSlice';
+
+export default configureStore({
+  reducer: {
+    // valueInThestore: reducerFromSlice
+    queryapi: fetchDataReducer,
+  },
+});
